+++ libs/tfest/src/TMatchingPair.cpp
@@ -7,17 +7,18 @@
    | Released under BSD License. See: https://www.mrpt.org/License          |
    +------------------------------------------------------------------------+ */
 
-#include "tfest-precomp.h"  // Precompiled headers
-
+#include "tfest-precomp.h"	// Precompiled headers
+//
 #include <mrpt/core/format.h>
 #include <mrpt/poses/CPose2D.h>
 #include <mrpt/system/os.h>
 #include <mrpt/tfest/TMatchingPair.h>
+
 #include <cmath>
 #include <cstdio>
 #include <fstream>
 #include <iostream>
-#include <numeric>  // accumulate()
+#include <numeric>	// accumulate()
 
 using namespace mrpt;
 using namespace mrpt::math;
@@ -26,7 +27,8 @@ using namespace mrpt::poses;
 using namespace mrpt::system;
 using namespace std;
 
-void TMatchingPairList::dumpToFile(const std::string& fileName) const
+template <typename T>
+void TMatchingPairListTempl<T>::dumpToFile(const std::string& fileName) const
 {
 	std::ofstream f(fileName);
 	ASSERT_(f.is_open());
@@ -39,7 +41,9 @@ void TMatchingPairList::dumpToFile(const std::string& fileName) const
 	}
 }
 
-void TMatchingPairList::saveAsMATLABScript(const std::string& filName) const
+template <typename T>
+void TMatchingPairListTempl<T>::saveAsMATLABScript(
+	const std::string& filName) const
 {
 	FILE* f = os::fopen(filName.c_str(), "wt");
 
@@ -49,7 +53,6 @@ void TMatchingPairList::saveAsMATLABScript(const std::string& filName) const
 	fprintf(
 		f, "%%  Before calling this script, define the color of lines, eg:\n");
 	fprintf(f, "%%     colorLines=[1 1 1]");
-	fprintf(f, "%%               J.L. Blanco (C) 2005-2012 \n");
 	fprintf(f, "%% ----------------------------------------------------\n\n");
 
 	fprintf(f, "axis equal; hold on;\n");
@@ -67,113 +70,82 @@ void TMatchingPairList::saveAsMATLABScript(const std::string& filName) const
 	os::fclose(f);
 }
 
-/*---------------------------------------------------------------
-						indexOtherMapHasCorrespondence
-  ---------------------------------------------------------------*/
-bool TMatchingPairList::indexOtherMapHasCorrespondence(size_t idx) const
+template <typename T>
+bool TMatchingPairListTempl<T>::indexOtherMapHasCorrespondence(size_t idx) const
 {
 	for (const auto& it : *this)
-	{
 		if (it.other_idx == idx) return true;
-	}
 	return false;
 }
 
-bool mrpt::tfest::operator<(const TMatchingPair& a, const TMatchingPair& b)
+template <typename T>
+T TMatchingPairListTempl<T>::overallSquareError(const CPose2D& q) const
 {
-	if (a.this_idx == b.this_idx)
-		return (a.this_idx < b.this_idx);
-	else
-		return (a.other_idx < b.other_idx);
-}
-
-bool mrpt::tfest::operator==(const TMatchingPair& a, const TMatchingPair& b)
-{
-	return (a.this_idx == b.this_idx) && (a.other_idx == b.other_idx);
-}
-
-bool mrpt::tfest::operator==(
-	const TMatchingPairList& a, const TMatchingPairList& b)
-{
-	if (a.size() != b.size()) return false;
-	for (auto it1 = a.begin(), it2 = b.begin(); it1 != a.end(); ++it1, ++it2)
-		if (!((*it1) == (*it2))) return false;
-	return true;
-}
-
-float TMatchingPairList::overallSquareError(const CPose2D& q) const
-{
-	vector<float> errs(size());
+	vector<T> errs(base_t::size());
 	squareErrorVector(q, errs);
-	return std::accumulate(errs.begin(), errs.end(), .0f);
+	return std::accumulate(errs.begin(), errs.end(), T(0));
 }
 
-float TMatchingPairList::overallSquareErrorAndPoints(
-	const CPose2D& q, vector<float>& xs, vector<float>& ys) const
+template <typename T>
+T TMatchingPairListTempl<T>::overallSquareErrorAndPoints(
+	const CPose2D& q, vector<T>& xs, vector<T>& ys) const
 {
-	vector<float> errs(size());
+	vector<T> errs(base_t::size());
 	squareErrorVector(q, errs, xs, ys);
-	return std::accumulate(errs.begin(), errs.end(), .0f);
+	return std::accumulate(errs.begin(), errs.end(), T(0));
 }
 
-/*---------------------------------------------------------------
-					TMatchingPairList::contains
-  ---------------------------------------------------------------*/
-bool TMatchingPairList::contains(const TMatchingPair& p) const
+template <typename T>
+bool TMatchingPairListTempl<T>::contains(const TMatchingPairTempl<T>& p) const
 {
 	for (const auto& corresp : *this)
 		if (corresp == p) return true;
 	return false;
 }
 
-/*---------------------------------------------------------------
-						squareErrorVector
-  ---------------------------------------------------------------*/
-void TMatchingPairList::squareErrorVector(
-	const CPose2D& q, vector<float>& out_sqErrs) const
+template <typename T>
+void TMatchingPairListTempl<T>::squareErrorVector(
+	const CPose2D& q, vector<T>& out_sqErrs) const
 {
-	out_sqErrs.resize(size());
+	out_sqErrs.resize(base_t::size());
 	// *    \f[ e_i = | x_{this} -  q \oplus x_{other}  |^2 \f]
 
-	const float ccos = d2f(std::cos(q.phi()));
-	const float csin = d2f(std::sin(q.phi()));
-	const float qx = d2f(q.x());
-	const float qy = d2f(q.y());
+	const T ccos = static_cast<T>(std::cos(q.phi()));
+	const T csin = static_cast<T>(std::sin(q.phi()));
+	const T qx = static_cast<T>(q.x());
+	const T qy = static_cast<T>(q.y());
 
-	const_iterator corresp;
-	vector<float>::iterator e_i;
-	for (corresp = begin(), e_i = out_sqErrs.begin(); corresp != end();
-		 ++corresp, ++e_i)
+	typename base_t::const_iterator corresp;
+	typename vector<T>::iterator e_i;
+	for (corresp = base_t::begin(), e_i = out_sqErrs.begin();
+		 corresp != base_t::end(); ++corresp, ++e_i)
 	{
-		float xx = qx + ccos * corresp->other_x - csin * corresp->other_y;
-		float yy = qy + csin * corresp->other_x + ccos * corresp->other_y;
+		T xx = qx + ccos * corresp->other_x - csin * corresp->other_y;
+		T yy = qy + csin * corresp->other_x + ccos * corresp->other_y;
 		*e_i = square(corresp->this_x - xx) + square(corresp->this_y - yy);
 	}
 }
 
-/*---------------------------------------------------------------
-						squareErrorVector
-  ---------------------------------------------------------------*/
-void TMatchingPairList::squareErrorVector(
-	const CPose2D& q, vector<float>& out_sqErrs, vector<float>& xs,
-	vector<float>& ys) const
+template <typename T>
+void TMatchingPairListTempl<T>::squareErrorVector(
+	const CPose2D& q, vector<T>& out_sqErrs, vector<T>& xs, vector<T>& ys) const
 {
-	out_sqErrs.resize(size());
-	xs.resize(size());
-	ys.resize(size());
+	out_sqErrs.resize(base_t::size());
+	xs.resize(base_t::size());
+	ys.resize(base_t::size());
 
 	// *    \f[ e_i = | x_{this} -  q \oplus x_{other}  |^2 \f]
 
-	const float ccos = d2f(cos(q.phi()));
-	const float csin = d2f(sin(q.phi()));
-	const float qx = d2f(q.x());
-	const float qy = d2f(q.y());
+	const T ccos = static_cast<T>(std::cos(q.phi()));
+	const T csin = static_cast<T>(std::sin(q.phi()));
+	const T qx = static_cast<T>(q.x());
+	const T qy = static_cast<T>(q.y());
 
-	const_iterator corresp;
-	vector<float>::iterator e_i, xx, yy;
-	for (corresp = begin(), e_i = out_sqErrs.begin(), xx = xs.begin(),
+	typename base_t::const_iterator corresp;
+	typename vector<T>::iterator e_i, xx, yy;
+	for (corresp = base_t::begin(), e_i = out_sqErrs.begin(), xx = xs.begin(),
 		yy = ys.begin();
-		 corresp != end(); ++corresp, ++e_i, ++xx, ++yy)
+		 corresp != base_t::end(); ++corresp, ++e_i, ++xx, ++yy)
 	{
 		*xx = qx + ccos * corresp->other_x - csin * corresp->other_y;
 		*yy = qy + csin * corresp->other_x + ccos * corresp->other_y;
@@ -181,10 +153,13 @@ void TMatchingPairList::squareErrorVector(
 	}
 }
 
-void TMatchingPairList::filterUniqueRobustPairs(
+template <typename T>
+void TMatchingPairListTempl<T>::filterUniqueRobustPairs(
 	const size_t num_elements_this_map,
-	TMatchingPairList& out_filtered_list) const
+	TMatchingPairListTempl<T>& out_filtered_list) const
 {
+	using TMatchingPairConstPtr = TMatchingPairTempl<T> const*;
+
 	std::vector<TMatchingPairConstPtr> bestMatchForThisMap(
 		num_elements_this_map, TMatchingPairConstPtr(nullptr));
 	out_filtered_list.clear();
@@ -208,18 +183,27 @@ void TMatchingPairList::filterUniqueRobustPairs(
 	for (auto& c : *this)
 	{
 		if (bestMatchForThisMap[c.this_idx] == &c)
-			out_filtered_list.push_back(c);  // Add to the output
+			out_filtered_list.push_back(c);	 // Add to the output
 	}
 }
 
-std::ostream& mrpt::tfest::operator<<(
-	std::ostream& o, const mrpt::tfest::TMatchingPair& pair)
+template <typename T>
+void TMatchingPairTempl<T>::print(std::ostream& o) const
 {
-	o << "[" << pair.this_idx << "->" << pair.other_idx << "]"
+	o << "[" << this_idx << "->" << other_idx << "]"
 	  << ": "
-	  << "(" << pair.this_x << "," << pair.this_y << "," << pair.this_z << ")"
+	  << "(" << this_x << "," << this_y << "," << this_z << ")"
 	  << " -> "
-	  << "(" << pair.other_x << "," << pair.other_y << "," << pair.other_z
-	  << ")";
-	return o;
+	  << "(" << other_x << "," << other_y << "," << other_z << ")";
 }
+
+// Explicit instantations:
+namespace mrpt::tfest
+{
+template class TMatchingPairListTempl<float>;
+template class TMatchingPairListTempl<double>;
+
+template class TMatchingPairTempl<float>;
+template class TMatchingPairTempl<double>;
+
+}  // namespace mrpt::tfest